+++ src/main/java/org/jdsystem/studentinfobackend/service/IStudentService.java
@@ -7,4 +7,5 @@ import java.util.List;
 public interface IStudentService {
     Student addStudent(Student student);
     List<Student> getStudents();
+    void deleteStudent(Long id);
 }
