--- src/main/java/org/jdsystem/studentinfobackend/sevice/IStudentService.java
@@ -1,9 +1,0 @@
-package org.jdsystem.studentinfobackend.sevice;
-
-import org.jdsystem.studentinfobackend.model.Student;
-
-import java.util.List;
-
-public interface IStudentService {
-    void deleteStudent(Long id);
-}
